+++ cub/test/catch2_test_device_scan_env.cu
@@ -68,7 +68,7 @@ TEST_CASE("Device scan exclusive scan works with default environment", "[scan][d
   REQUIRE(d_out[1] == (init + value_t{1}));
 
   // Make sure we use default tuning
-  REQUIRE(d_block_size[0] == target_block_size);
+  REQUIRE(d_block_size[0] == static_cast<unsigned int>(target_block_size));
 }
 
 TEST_CASE("Device scan exclusive scan with FutureValue works with default environment", "[scan][device]")
@@ -138,7 +138,7 @@ using block_sizes = c2h::type_list<cuda::std::integral_constant<int, 32>, cuda::
 
 C2H_TEST("Device scan exclusive-scan can be tuned", "[scan][device]", block_sizes)
 {
-  constexpr int target_block_size = c2h::get<0, TestType>::value;
+  constexpr unsigned int target_block_size = c2h::get<0, TestType>::value;
   c2h::device_vector<unsigned int> d_block_size(1);
   block_size_check_t block_size_check{thrust::raw_pointer_cast(d_block_size.data())};
 
@@ -158,7 +158,7 @@ C2H_TEST("Device scan exclusive-scan can be tuned", "[scan][device]", block_size
 
 C2H_TEST("Device scan exclusive-sum can be tuned", "[scan][device]", block_sizes)
 {
-  constexpr int target_block_size = c2h::get<0, TestType>::value;
+  constexpr unsigned int target_block_size = c2h::get<0, TestType>::value;
 
   auto num_items = target_block_size;
   c2h::device_vector<unsigned int> d_block_size(1, 0);
@@ -217,12 +217,12 @@ TEST_CASE("Device scan inclusive-scan works with default environment", "[scan][d
   REQUIRE(d_out[1] == d_in[0] + d_in[1]);
 
   // Make sure we use default tuning
-  REQUIRE(d_block_size[0] == target_block_size);
+  REQUIRE(d_block_size[0] == static_cast<unsigned int>(target_block_size));
 }
 
 C2H_TEST("Device scan inclusive-scan can be tuned", "[scan][device]", block_sizes)
 {
-  constexpr int target_block_size = c2h::get<0, TestType>::value;
+  constexpr unsigned int target_block_size = c2h::get<0, TestType>::value;
   c2h::device_vector<unsigned int> d_block_size(1);
   block_size_check_t block_size_check{thrust::raw_pointer_cast(d_block_size.data())};
 
@@ -257,7 +257,7 @@ TEST_CASE("Device scan inclusive-scan-init works with default environment", "[sc
 
 C2H_TEST("Device scan inclusive-scan-init can be tuned", "[scan][device]", block_sizes)
 {
-  constexpr int target_block_size = c2h::get<0, TestType>::value;
+  constexpr unsigned int target_block_size = c2h::get<0, TestType>::value;
   c2h::device_vector<unsigned int> d_block_size(1);
   block_size_check_t block_size_check{thrust::raw_pointer_cast(d_block_size.data())};
